+++ src/service1/pages/static/pages/index/js/hover_pong.js
@@ -2,13 +2,17 @@ document.addEventListener("DOMContentLoaded", function() {
     const indexPong = document.querySelector('.index-pong');
     const indexSudoku = document.querySelector('.index-sudoku');
 
-    indexPong.addEventListener('mouseenter', function() {
-        indexPong.classList.add('scaled-pong');
-        indexSudoku.classList.add('shrunken-pong');
-    });
 
-    indexPong.addEventListener('mouseleave', function() {
-        indexPong.classList.remove('scaled-pong');
-        indexSudoku.classList.remove('shrunken-pong');
-    });
+    if (indexSudoku && indexPong)
+    {
+        indexPong.addEventListener('mouseenter', function() {
+            indexPong.classList.add('scaled-pong');
+            indexSudoku.classList.add('shrunken-pong');
+        });
+    
+        indexPong.addEventListener('mouseleave', function() {
+            indexPong.classList.remove('scaled-pong');
+            indexSudoku.classList.remove('shrunken-pong');
+        });
+    }
 });
