+++ src/service1/pages/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render
 from django.http import HttpResponse
 
 def home(request):
-    return render(request, 'pages/base.html')
+    return render(request, 'pages/index.html')
 
 def login(request):
     return render(request, 'pages/login.html')
